Simplify AR mixin
- Using `@Group` to consolidate v1.7.0 and v2.0.0 into one mixin
