+++ src/main/java/org/dimdev/jeid/core/JEIDMixinLoader.java
@@ -1,12 +1,11 @@
 package org.dimdev.jeid.core;
 
-import net.minecraftforge.fml.common.Loader;
-import net.minecraftforge.fml.common.ModContainer;
-import zone.rong.mixinbooter.ILateMixinLoader;
-
 import java.util.ArrayList;
 import java.util.List;
 
+import net.minecraftforge.fml.common.Loader;
+
+import zone.rong.mixinbooter.ILateMixinLoader;
 
 public class JEIDMixinLoader implements ILateMixinLoader {
     public List<String> getMixinConfigs() {
@@ -21,12 +20,7 @@ public class JEIDMixinLoader implements ILateMixinLoader {
             configs.add("mixins.jeid.abyssalcraft.json");
         }
         if (Loader.isModLoaded("advancedrocketry")) {
-            String version = Loader.instance().getIndexedModList().get("advancedrocketry").getVersion();
-            if (version.split("-")[1].contains(".") && Integer.parseInt(version.split("-")[1].split("\\.")[0]) >= 2) {
-                configs.add("mixins.jeid.advancedrocketry.v2_0_0.json");
-            } else {
-                configs.add("mixins.jeid.advancedrocketry.v1_7_0.json");
-            }
+            configs.add("mixins.jeid.advancedrocketry.json");
         }
         if (Loader.isModLoaded("atum")) {
             configs.add("mixins.jeid.atum.json");
