+++ src/main/resources/mixins.jeid.advancedrocketry.json
@@ -5,7 +5,7 @@
   "minVersion": "0.8",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "v2_0_0.MixinBiomeHandler",
+    "MixinBiomeHandler",
     "MixinPacketBiomeIDChange"
   ]
 }
