--- src/main/resources/mixins.jeid.advancedrocketry.v1_7_0.json
@@ -1,11 +1,0 @@
-{
-  "package": "org.dimdev.jeid.mixin.modsupport.advancedrocketry",
-  "required": true,
-  "refmap": "mixins.jeid.refmap.json",
-  "minVersion": "0.8",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [
-    "v1_7_0.MixinBiomeHandler",
-    "MixinPacketBiomeIDChange"
-  ]
-}
